requirements.txt
aniso8601==9.0.1
antiorm==1.2.1
automium==0.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==7.1.2
db==0.1.1
Deprecated==1.2.13
elastic-transport==8.0.1
elasticsearch==8.0.0
Flask==2.0.3
Flask-JWT-Extended==4.3.1
flask-marshmallow==0.14.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
legacy==0.1.6
listener==2.0.0a1
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
packaging==21.3
pamqp==2.3.0
pika==1.2.0
pika-stubs==0.1.3
PyJWT==2.3.0
PyMySQL==1.0.2
pyparsing==3.0.7
pytz==2021.3
rabbitpy==2.0.1
redis==4.1.4
requests==2.27.1
six==1.16.0
SQLAlchemy==1.4.31
urllib3==1.26.8
urlparser==0.1.2
Werkzeug==2.0.3
wrapt==1.13.3
